test to integrate validator config
