+++ clinicadl/trainer/tasks_utils.py
@@ -207,175 +207,6 @@ def evaluation_metrics(network_task: Union[str, Task]):
         raise ValueError("Unknown network task")
 
 
-def test(
-    mode: str,
-    metrics_module: MetricModule,
-    n_classes: int,
-    network_task,
-    model: Network,
-    dataloader: DataLoader,
-    criterion: _Loss,
-    use_labels: bool = True,
-    amp: bool = False,
-    report_ci=False,
-) -> Tuple[pd.DataFrame, Dict[str, float]]:
-    """
-    Computes the predictions and evaluation metrics.
-
-    Parameters
-    ----------
-    model: Network
-        The model trained.
-    dataloader: DataLoader
-        Wrapper of a CapsDataset.
-    criterion:  _Loss
-        Function to calculate the loss.
-    use_labels: bool
-        If True the true_label will be written in output DataFrame
-        and metrics dict will be created.
-    amp: bool
-        If True, enables Pytorch's automatic mixed precision.
-
-    Returns
-    -------
-        the results and metrics on the image level.
-    """
-    model.eval()
-    dataloader.dataset.eval()
-
-    results_df = pd.DataFrame(columns=columns(network_task, mode, n_classes))
-    total_loss = {}
-    with torch.no_grad():
-        for i, data in enumerate(dataloader):
-            # initialize the loss list to save the loss components
-            with autocast("cuda", enabled=amp):
-                outputs, loss_dict = model(data, criterion, use_labels=use_labels)
-
-            if i == 0:
-                for loss_component in loss_dict.keys():
-                    total_loss[loss_component] = 0
-            for loss_component in total_loss.keys():
-                total_loss[loss_component] += loss_dict[loss_component].float()
-
-            # Generate detailed DataFrame
-            for idx in range(len(data["participant_id"])):
-                row = generate_test_row(
-                    network_task,
-                    mode,
-                    metrics_module,
-                    n_classes,
-                    idx,
-                    data,
-                    outputs.float(),
-                )
-                row_df = pd.DataFrame(
-                    row, columns=columns(network_task, mode, n_classes)
-                )
-                results_df = pd.concat([results_df, row_df])
-
-            del outputs, loss_dict
-    dataframes = [None] * dist.get_world_size()
-    dist.gather_object(results_df, dataframes if dist.get_rank() == 0 else None, dst=0)
-    if dist.get_rank() == 0:
-        results_df = pd.concat(dataframes)
-    del dataframes
-    results_df.reset_index(inplace=True, drop=True)
-
-    if not use_labels:
-        metrics_dict = None
-    else:
-        metrics_dict = compute_metrics(
-            network_task, results_df, metrics_module, report_ci=report_ci
-        )
-        for loss_component in total_loss.keys():
-            dist.reduce(total_loss[loss_component], dst=0)
-            loss_value = total_loss[loss_component].item() / cluster.world_size
-
-            if report_ci:
-                metrics_dict["Metric_names"].append(loss_component)
-                metrics_dict["Metric_values"].append(loss_value)
-                metrics_dict["Lower_CI"].append("N/A")
-                metrics_dict["Upper_CI"].append("N/A")
-                metrics_dict["SE"].append("N/A")
-
-            else:
-                metrics_dict[loss_component] = loss_value
-
-    torch.cuda.empty_cache()
-
-    return results_df, metrics_dict
-
-
-def test_da(
-    mode: str,
-    metrics_module: MetricModule,
-    n_classes: int,
-    network_task: Union[str, Task],
-    model: Network,
-    dataloader: DataLoader,
-    criterion: _Loss,
-    alpha: float = 0,
-    use_labels: bool = True,
-    target: bool = True,
-    report_ci=False,
-) -> Tuple[pd.DataFrame, Dict[str, float]]:
-    """
-    Computes the predictions and evaluation metrics.
-
-    Args:
-        model: the model trained.
-        dataloader: wrapper of a CapsDataset.
-        criterion: function to calculate the loss.
-        use_labels: If True the true_label will be written in output DataFrame
-            and metrics dict will be created.
-    Returns:
-        the results and metrics on the image level.
-    """
-    model.eval()
-    dataloader.dataset.eval()
-    results_df = pd.DataFrame(columns=columns(network_task, mode, n_classes))
-    total_loss = 0
-    with torch.no_grad():
-        for i, data in enumerate(dataloader):
-            outputs, loss_dict = model.compute_outputs_and_loss_test(
-                data, criterion, alpha, target
-            )
-            total_loss += loss_dict["loss"].item()
-
-            # Generate detailed DataFrame
-            for idx in range(len(data["participant_id"])):
-                row = generate_test_row(
-                    network_task, mode, metrics_module, n_classes, idx, data, outputs
-                )
-                row_df = pd.DataFrame(
-                    row, columns=columns(network_task, mode, n_classes)
-                )
-                results_df = pd.concat([results_df, row_df])
-
-            del outputs, loss_dict
-        results_df.reset_index(inplace=True, drop=True)
-
-    if not use_labels:
-        metrics_dict = None
-    else:
-        metrics_dict = compute_metrics(
-            network_task, results_df, metrics_module, report_ci=report_ci
-        )
-        if report_ci:
-            metrics_dict["Metric_names"].append("loss")
-            metrics_dict["Metric_values"].append(total_loss)
-            metrics_dict["Lower_CI"].append("N/A")
-            metrics_dict["Upper_CI"].append("N/A")
-            metrics_dict["SE"].append("N/A")
-
-        else:
-            metrics_dict["loss"] = total_loss
-
-    torch.cuda.empty_cache()
-
-    return results_df, metrics_dict
-
-
 def columns(network_task: Union[str, Task], mode: str, n_classes: Optional[int] = None):
     """
     List of the columns' names in the TSV file containing the predictions.
