+++ clinicadl/validator/validator.py
@@ -1,19 +1,31 @@
 from logging import getLogger
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
+import numpy as np
+import pandas as pd
 import torch
 import torch.distributed as dist
 from torch.amp import autocast
+from torch.nn.modules.loss import _Loss
+from torch.utils.data import DataLoader
 
 from clinicadl.maps_manager.maps_manager import MapsManager
+from clinicadl.metrics.metric_module import MetricModule
 from clinicadl.metrics.utils import find_selection_metrics
-from clinicadl.trainer.tasks_utils import (
-    test,
-    test_da,
-)
+from clinicadl.network.network import Network
+from clinicadl.trainer.tasks_utils import columns, compute_metrics, generate_test_row
 from clinicadl.utils import cluster
 from clinicadl.utils.computational.ddp import DDP, init_ddp
+from clinicadl.utils.enum import (
+    ClassificationLoss,
+    ClassificationMetric,
+    ReconstructionLoss,
+    ReconstructionMetric,
+    RegressionLoss,
+    RegressionMetric,
+    Task,
+)
 from clinicadl.utils.exceptions import (
     ClinicaDLArgumentError,
     ClinicaDLConfigurationError,
@@ -41,6 +53,183 @@ class Validator:
         # maps_manager.mode = mode
 
         print("test")
+
+    def test(
+        self,
+        mode: str,
+        metrics_module: MetricModule,
+        n_classes: int,
+        network_task,
+        model: Network,
+        dataloader: DataLoader,
+        criterion: _Loss,
+        use_labels: bool = True,
+        amp: bool = False,
+        report_ci=False,
+    ) -> Tuple[pd.DataFrame, Dict[str, float]]:
+        """
+        Computes the predictions and evaluation metrics.
+
+        Parameters
+        ----------
+        model: Network
+            The model trained.
+        dataloader: DataLoader
+            Wrapper of a CapsDataset.
+        criterion:  _Loss
+            Function to calculate the loss.
+        use_labels: bool
+            If True the true_label will be written in output DataFrame
+            and metrics dict will be created.
+        amp: bool
+            If True, enables Pytorch's automatic mixed precision.
+
+        Returns
+        -------
+            the results and metrics on the image level.
+        """
+        model.eval()
+        dataloader.dataset.eval()
+
+        results_df = pd.DataFrame(columns=columns(network_task, mode, n_classes))
+        total_loss = {}
+        with torch.no_grad():
+            for i, data in enumerate(dataloader):
+                # initialize the loss list to save the loss components
+                with autocast("cuda", enabled=amp):
+                    outputs, loss_dict = model(data, criterion, use_labels=use_labels)
+
+                if i == 0:
+                    for loss_component in loss_dict.keys():
+                        total_loss[loss_component] = 0
+                for loss_component in total_loss.keys():
+                    total_loss[loss_component] += loss_dict[loss_component].float()
+
+                # Generate detailed DataFrame
+                for idx in range(len(data["participant_id"])):
+                    row = generate_test_row(
+                        network_task,
+                        mode,
+                        metrics_module,
+                        n_classes,
+                        idx,
+                        data,
+                        outputs.float(),
+                    )
+                    row_df = pd.DataFrame(
+                        row, columns=columns(network_task, mode, n_classes)
+                    )
+                    results_df = pd.concat([results_df, row_df])
+
+                del outputs, loss_dict
+        dataframes = [None] * dist.get_world_size()
+        dist.gather_object(
+            results_df, dataframes if dist.get_rank() == 0 else None, dst=0
+        )
+        if dist.get_rank() == 0:
+            results_df = pd.concat(dataframes)
+        del dataframes
+        results_df.reset_index(inplace=True, drop=True)
+
+        if not use_labels:
+            metrics_dict = None
+        else:
+            metrics_dict = compute_metrics(
+                network_task, results_df, metrics_module, report_ci=report_ci
+            )
+            for loss_component in total_loss.keys():
+                dist.reduce(total_loss[loss_component], dst=0)
+                loss_value = total_loss[loss_component].item() / cluster.world_size
+
+                if report_ci:
+                    metrics_dict["Metric_names"].append(loss_component)
+                    metrics_dict["Metric_values"].append(loss_value)
+                    metrics_dict["Lower_CI"].append("N/A")
+                    metrics_dict["Upper_CI"].append("N/A")
+                    metrics_dict["SE"].append("N/A")
+
+                else:
+                    metrics_dict[loss_component] = loss_value
+
+        torch.cuda.empty_cache()
+
+        return results_df, metrics_dict
+
+    def test_da(
+        self,
+        mode: str,
+        metrics_module: MetricModule,
+        n_classes: int,
+        network_task: Union[str, Task],
+        model: Network,
+        dataloader: DataLoader,
+        criterion: _Loss,
+        alpha: float = 0,
+        use_labels: bool = True,
+        target: bool = True,
+        report_ci=False,
+    ) -> Tuple[pd.DataFrame, Dict[str, float]]:
+        """
+        Computes the predictions and evaluation metrics.
+
+        Args:
+            model: the model trained.
+            dataloader: wrapper of a CapsDataset.
+            criterion: function to calculate the loss.
+            use_labels: If True the true_label will be written in output DataFrame
+                and metrics dict will be created.
+        Returns:
+            the results and metrics on the image level.
+        """
+        model.eval()
+        dataloader.dataset.eval()
+        results_df = pd.DataFrame(columns=columns(network_task, mode, n_classes))
+        total_loss = 0
+        with torch.no_grad():
+            for i, data in enumerate(dataloader):
+                outputs, loss_dict = model.compute_outputs_and_loss_test(
+                    data, criterion, alpha, target
+                )
+                total_loss += loss_dict["loss"].item()
+
+                # Generate detailed DataFrame
+                for idx in range(len(data["participant_id"])):
+                    row = generate_test_row(
+                        network_task,
+                        mode,
+                        metrics_module,
+                        n_classes,
+                        idx,
+                        data,
+                        outputs,
+                    )
+                    row_df = pd.DataFrame(
+                        row, columns=columns(network_task, mode, n_classes)
+                    )
+                    results_df = pd.concat([results_df, row_df])
+
+                del outputs, loss_dict
+            results_df.reset_index(inplace=True, drop=True)
+
+        if not use_labels:
+            metrics_dict = None
+        else:
+            metrics_dict = compute_metrics(
+                network_task, results_df, metrics_module, report_ci=report_ci
+            )
+            if report_ci:
+                metrics_dict["Metric_names"].append("loss")
+                metrics_dict["Metric_values"].append(total_loss)
+                metrics_dict["Lower_CI"].append("N/A")
+                metrics_dict["Upper_CI"].append("N/A")
+                metrics_dict["SE"].append("N/A")
+
+            else:
+                metrics_dict["loss"] = total_loss
+
+        torch.cuda.empty_cache()
+
+        return results_df, metrics_dict
 
     def _test_loader(
         self,
@@ -99,7 +288,7 @@ class Validator:
                 amp=maps_manager.amp,
             )
 
-            prediction_df, metrics = test(
+            prediction_df, metrics = self.test(
                 mode=maps_manager.mode,
                 metrics_module=maps_manager.metrics_module,
                 n_classes=maps_manager.n_classes,
@@ -183,7 +372,7 @@ class Validator:
                 gpu=gpu,
                 network=network,
             )
-            prediction_df, metrics = test_da(
+            prediction_df, metrics = self.test_da(
                 network_task=maps_manager.network_task,
                 model=model,
                 dataloader=dataloader,
